pyproject.toml
[project]
name = "ai-nexus"
version = "0.1.0"
description = "ai-nexus agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph~=0.4.0",
    "langchain~=0.3.8",
    "langchain-core~=0.3.8",
    "python-dotenv~=1.0.1",
    "langgraph-sdk~=0.1.32",
    "langchain-google-genai~=2.1.4",
    "pytest-asyncio~=0.26.0",
    "pygithub~=2.6.1",
    "langchain-community~=0.3.23",
    "langmem~=0.0.25",
    "termcolor~=3.1.0",
    "agentevals~=0.0.7",
    "fastapi~=0.115.0",
    "uvicorn[standard]~=0.26.0",
    "httpx~=0.25.0",
    "PyJWT~=2.9.0",
    "pydantic-settings~=2.4.0",
]
optional-dependencies.dev = [
    "mypy~=1.11.1",
    "ruff~=0.6.1",
    "pickpack>=2.0.0",
    "anytree>=2.13.0",
    "pytest~=8.3.5",
    "pytest-asyncio~=0.26.0",
    "pytest-dotenv~=0.5.2",
    "langgraph-cli[inmem]~=0.2.7",
    "openevals~=0.0.19",
    "debugpy~=1.8.14",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# Add new agents here: #1
[tool.setuptools]
packages = [
    "common",
    "agent_template",
    "code_reviewer",
    "orchestrator",
    "architect",
    "coder",
    "tester",
    "requirement_gatherer",
    "task_manager",
    "scenarios",
    "pr_memory_updater",
    "event_listener",
]
[tool.setuptools.package-dir]
"common" = "src/common"
"agent_template" = "src/agent_template"
"code_reviewer" = "src/code_reviewer"
"langgraph.templates.agent_template" = "src/agent_template"
"orchestrator" = "src/orchestrator"
"langgraph.templates.orchestrator" = "src/orchestrator"
"tester" = "src/tester"
"grumpy" = "src/grumpy"
"architect" = "src/architect"
"langgraph.templates.grumpy" = "src/grumpy"
"requirement_gatherer" = "src/requirement_gatherer"
"langgraph.templates.requirement_gatherer" = "src/requirement_gatherer"
"coder" = "src/coder"
"langgraph.templates.coder" = "src/coder"
"task_manager" = "src/task_manager"
"pr_memory_updater" = "src/pr_memory_updater"
"langgraph.templates.task_manager" = "src/task_manager"
"scenarios" = "tests/scenarios"
"event_listener" = "src/event_listener"


[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
include = ["*.py", "*.pyi", "*.ipynb"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP", "T201"]
"src/demo/*" = ["D", "UP", "T201"]
"src/orchestrator/{graph,test}.py" = ["T201"]
"src/requirement_gatherer/graph.py" = ["T201"]
"ntbk/*" = ["D", "UP", "T201"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_errors = false

[tool.codespell]
skip = "node_modules"
